https://sourceforge.net/tracker2/?func=detail&aid=2488108&group_id=74339&atid=540676
Remove the distance-to-vertex check on triangles in getCollisionPoint(), since it was erroneously rejecting close collisions with triangles with distant vertices.  Added a unit test (verifying that it selects the wrong triangle with unpatched code).

Also made a minor fix to the drawRectOutline() test, so that the driver used matches the control image in SVN.